Reduce multiple classes into a single decorator
I refactored the social media link classes that previously inherited
from `ExternalLink::Base` into a single `ExternalLink` decorator class.
This change eliminates the need for separate classes
for each social media platform, ensuring codebase consistency.
Additionally, I introduced a hash map for easy mapping of link types
to their respective social media icons.
Now, when providing input, we only need to specify the URL
and link type; there is no longer a need to provide the title.
The ability to provide a title for overwriting
the default translation value remains intact.
