--- app/models/external_link/blog.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Blog < Base
-    private
-
-    def image_url
-      '/icons/common/blog/blogger.svg'
-    end
-  end
-end
