--- app/models/external_link/facebook.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Facebook < Base
-    private
-
-    def image_url
-      '/icons/social-media/facebook/facebook-original.svg'
-    end
-  end
-end
