--- app/models/external_link/github.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Github < Base
-    private
-
-    def image_url
-      '/icons/technology/github/github-original.svg'
-    end
-  end
-end
