--- app/models/external_link/instagram.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Instagram < Base
-    private
-
-    def image_url
-      '/icons/social-media/instagram/instagram-original.svg'
-    end
-  end
-end
