--- app/models/external_link/linkedin.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Linkedin < Base
-    private
-
-    def image_url
-      '/icons/social-media/linkedin/linkedin-original.svg'
-    end
-  end
-end
