--- app/models/external_link/twitter.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Twitter < Base
-    private
-
-    def image_url
-      '/icons/social-media/twitter/twitter-original.svg'
-    end
-  end
-end
