--- app/models/external_link/vimeo.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Vimeo < Base
-    private
-
-    def image_url
-      '/icons/social-media/vimeo/vimeo-original.svg'
-    end
-  end
-end
