--- app/models/external_link/website.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Website < Blog
-    private
-
-    def image_url
-      '/icons/common/website/website.svg'
-    end
-  end
-end
