--- app/models/external_link/youtube.rb
@@ -1,9 +1,0 @@
-module ExternalLink
-  class Youtube < Base
-    private
-
-    def image_url
-      '/icons/social-media/youtube/youtube-original.svg'
-    end
-  end
-end
